Revert "Merge remote-tracking branch 'upstream/master' into GulagDisposals"
This reverts commit fb0017fb3593cfd2fe154973628852209f60d097, reversing
changes made to 208659efe149a059bb39e15ffd2173a3ba711b78.

+++ code/game/objects/items/implants/implant_stealth.dm
@@ -17,15 +17,19 @@
 	move_speed_multiplier = 0.5
 	enable_door_overlay = FALSE
 
+/obj/structure/closet/cardboard/agent/proc/go_invisible()
+	animate(src, , alpha = 0, time = 20)
+
 /obj/structure/closet/cardboard/agent/Initialize(mapload)
 	. = ..()
 	go_invisible()
 
-/obj/structure/closet/cardboard/agent/proc/go_invisible()
-	animate(src, alpha = 0, time = 20)
-
-/obj/structure/closet/cardboard/agent/after_open(mob/living/user)
+/obj/structure/closet/cardboard/agent/open(mob/living/user, force = FALSE)
 	. = ..()
+
+	if(!.)
+		return
+
 	qdel(src)
 
 /obj/structure/closet/cardboard/agent/process()
